+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.1.0
+
+* Adding Custom Lane Builder
+* Adding Custom Item Builder
+* Adding Support for children instead of app
+* Bug Fixes
+* Changing Layout Engine
+* Reset now defaults to initial defaults
+* Adding CrossAxisCount for wrapping rows
+* Support for dynamic sizes
+* Adding multiple examples
+* Changing AppBar to PreferredSizeWidget
+
 ## 0.0.4
 
 * Adding Support for Route Arguments